check migration price column

+++ database/migrations/2022_11_16_161752_create_plates_table.php
@@ -18,7 +18,7 @@ class CreatePlatesTable extends Migration
             $table->string('name');
             $table->string('img')->nullable();
             $table->text('description');
-            $table->smallInteger('price');
+            $table->double('price',4,2);
             $table->boolean('visible');
             $table->string('slug')->unique();
             $table->timestamps();
